requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
pillow==10.1.0
flask==3.1.2
pytest==8.4.2
